requirements.txt
fastapi
uvicorn
python-dotenv
openai
duckduckgo-search
pydantic
